requirements.txt
###### Requirements with Version Specifiers ######
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
oauth2client
chevron 